+++ ddl/user_group.sql
@@ -1,6 +1,6 @@
 CREATE TABLE USER_GROUP (
-  user_id int not null references APP_USER(id) on delete cascade,
-  group_id int not null references APP_GROUP(id) on delete cascade,
+  user_id bigint not null references APP_USER(id) on delete cascade,
+  group_id bigint not null references APP_GROUP(id) on delete cascade,
   created_at timestamp not null default current_timestamp,
   updated_at timestamp not null default current_timestamp,
   primary key (user_id, group_id)
